requirements.txt
pygame-ce
numpy
pymunk
setuptools
dotenv==0.9.9
invoke==2.2.0
python-dotenv==1.0.1
sphinx<9
sphinx-intl
sphinxcontrib.video
sphinx_autodoc_typehints
myst_parser
sphinxcontrib.mermaid
autoclasstoc
sphinx_toolbox
sphinx_book_theme